web/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0.0
Pillow>=10.0.0

